Refactor: move IPO to headers (enabling gtest)
